+++ package.json
@@ -24,13 +24,13 @@
     "bcrypt": "^5.1.1",
     "eslint": "^8.57.0",
     "js-yaml": "^4.1.0",
-    "ms": "^2.1.3",
+    "jsonwebtoken": "^9.0.2",
     "postcss": "^8.4.38",
     "postcss-load-config": "^4.0.2",
     "prettier": "^2.8.8",
     "prettier-plugin-svelte": "^2.10.1",
     "prettier-plugin-tailwindcss": "^0.2.8",
-    "sass": "^1.77.2",
+    "sass": "^1.77.4",
     "svelte": "^3.59.2",
     "svelte-check": "^2.10.3",
     "svelte-preprocess": "^4.10.7",
@@ -43,7 +43,7 @@
   "dependencies": {
     "@trpc/client": "^10.45.2",
     "@trpc/server": "^10.45.2",
-    "jsonwebtoken": "^9.0.2",
+    "ms": "^2.1.3",
     "svelte-icons-pack": "^2.1.1",
     "trpc-sveltekit": "^3.6.1",
     "zod": "^3.23.8"
