+++ src/components/drive/FileBar/MoveFile.svelte
@@ -7,7 +7,7 @@
   import { to_base64 } from '@tools/kry/gupta';
   import { toast } from '@tools/toast';
   import { client } from '@api/client';
-  import { ensure_jwt_status } from '@tools/drive/request';
+  import { ensure_jwt_status } from '@tools/auth_tools';
 
   $: lekh = $lekhAH.fileBar.MoveFile;
