+++ src/hooks.server.ts
@@ -5,7 +5,7 @@ import { router } from '@api/trpc_router';
 import { createTRPCHandle } from 'trpc-sveltekit';
 import { get_link, get_locale_from_url, get_pure_link } from '@tools/i18n';
 import { get_verified_id_token_info } from '@tools/jwt';
-import { AUTH_ID } from '@tools/drive/cookie_info';
+import { AUTH_ID } from '@tools/auth_tools';
 
 const handle_trpc = createTRPCHandle({ router, createContext });
 
@@ -15,9 +15,9 @@ export const handle: Handle = async ({ event, resolve }) => {
   }
 
   if (event.request.method === 'GET') {
-    const protected_routes = ['/drive'];
-    const LOGIN_URL = '/drive/login';
-    const DRIVE_URL = '/drive';
+    const protected_routes = ['/'];
+    const LOGIN_URL = '/login';
+    const DRIVE_URL = '/';
     const locale = get_locale_from_url(event.url.pathname);
     const URL = get_pure_link(event.url.pathname);
