+++ src/routes/[[lang]]/+page.server.ts
